add render.js
