+++ src/util.js
@@ -16,16 +16,6 @@ export const castDate = (date) => {
   return `${hours}:${minutes}`;
 };
 
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  AFTEREND: `afterend`,
-  BEFOREEND: `beforeend`
-};
-
-// export const formatEditDate = (date) => {
-//   return formatDateToICO(date).split(`T`).join(` `);
-// };
-
 export const formatEditDate = (date) => {
   return `${date.getDay()}/${date.getMonth()}/${date.getFullYear()} ${date.getHours()}:${date.getMinutes()}`;
 };
@@ -37,25 +27,4 @@ export const formatDateAttribute = (date) => {
 export const formatToTitleCase = (word) => {
   const firstLetter = word[0].toUpperCase();
   return `${firstLetter}${word.slice(1)}`;
-};
-
-export const createElement = (template) => {
-  const container = document.createElement(`div`);
-  container.innerHTML = template;
-
-  return container;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-    case RenderPosition.AFTEREND:
-      container.after(element);
-      break;
-  }
 };
